requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
python-dotenv==0.19.2
requests==2.27.1
urllib3==1.26.8
